+++ .idea/StretchTimer.iml
@@ -7,4 +7,14 @@
     <orderEntry type="jdk" jdkName="Python 3.10 (StretchTimer)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="TemplatesService">
+    <option name="TEMPLATE_FOLDERS">
+      <list>
+        <option value="$MODULE_DIR$/app/templates" />
+        <option value="$MODULE_DIR$/app/templates/accounts" />
+        <option value="$MODULE_DIR$/app/templates/includes" />
+        <option value="$MODULE_DIR$/app/templates/layouts" />
+      </list>
+    </option>
+  </component>
 </module>
