+++ app/templates/accounts/register.html
@@ -9,6 +9,6 @@
     <title>Document</title>
 </head>
 <body>
-<h1>Lets Go!</h1>
+<h1>Lets Go 2!</h1>
 </body>
 </html>
