+++ app/timer.py
@@ -1,1 +1,3 @@
 from datetime import date
+
+# TODO: Timer Back-end Logic
