+++ run.py
@@ -10,24 +10,10 @@ from flask_sqlalchemy import SQLAlchemy
 from werkzeug.security import generate_password_hash, check_password_hash
 from werkzeug.security import generate_password_hash, check_password_hash
 
+from app import app
 from app.forms import RegisterForm, LoginForm
-from app.tables import User
-from app.decorators import admin_only
 
-# from flask_ckeditor import CKEditor
-
-
-app = Flask(__name__)
-app.config['SECRET_KEY'] = '8BYkEfBA6O6donzWlSihBXox7C0sKR6b'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///stretch-timer.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-app.debug = True
-toolbar = DebugToolbarExtension(app)
-# ckeditor = CKEditor(app)
-Bootstrap(app)
-db = SQLAlchemy(app)
-login_manager = LoginManager()
-login_manager.init_app(app)
+# from app.tables import User
 
 
 # gravatar = Gravatar(app, size=100, rating='g', default='retro', force_default=False, force_lower=False,
@@ -35,63 +21,63 @@ login_manager.init_app(app)
 #                     base_url=None)
 
 
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(user_id)
+# @login_manager.user_loader
+# def load_user(user_id):
+#     return User.query.get(user_id)
 
 
-with app.app_context():
-    db.create_all()
+# with app.app_context():
+#     db.create_all()
 
 
-@app.route('/')
-def home():
-    # return render_template("index.html")
-    return '<p>Hello</p>'
+# @app.route('/')
+# def home():
+#     return render_template("app/index.html")
 
 
-@app.route('/register', methods=["GET", "POST"])
-def register():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        email = form.email.data
-        if User.query.filter_by(email=email).first():
-            flash("Email already exists. Try to log in.")
-            return redirect(url_for("login"))
-        password = generate_password_hash(form.password.data, salt_length=8)
-        name = form.name.data
-        new_user = User(email=email, password=password, name=name)
-        db.session.add(new_user)
-        db.session.commit()
-        login_user(User.query.filter_by(email=email).first())
-        return redirect(url_for("get_all_posts"))
-    return render_template("register.html", form=form)
-
-
-@app.route('/login', methods=["GET", "POST"])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        email = form.email.data
-        user = User.query.filter_by(email=email).first()
-        if not user:
-            flash("Unknown Email, Try Again.")
-            return render_template("login.html", form=form)
-        password = form.password.data
-        if not check_password_hash(user.password, password):
-            flash("Wrong Password, Try Again.")
-            return render_template("/app/templates/accounts/login.html", form=form)
-        login_user(user)
-        return redirect(url_for("get_all_posts"))
-    return render_template("login.html", form=form)
-
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('get_all_posts'))
+#
+# @app.route('/register', methods=["GET", "POST"])
+# def register():
+#     form = RegisterForm()
+#     if form.validate_on_submit():
+#         email = form.email.data
+#         if User.query.filter_by(email=email).first():
+#             flash("Email already exists. Try to log in.")
+#             return redirect(url_for("login"))
+#         password = generate_password_hash(form.password.data, salt_length=8)
+#         name = form.name.data
+#         new_user = User(email=email, password=password, name=name)
+#         db.session.add(new_user)
+#         db.session.commit()
+#         login_user(User.query.filter_by(email=email).first())
+#         return redirect(url_for("get_all_posts"))
+#     return render_template("register.html", form=form)
+#
+#
+# @app.route('/login', methods=["GET", "POST"])
+# def login():
+#     form = LoginForm()
+#     if form.validate_on_submit():
+#         email = form.email.data
+#         user = User.query.filter_by(email=email).first()
+#         if not user:
+#             flash("Unknown Email, Try Again.")
+#             return render_template("login.html", form=form)
+#         password = form.password.data
+#         if not check_password_hash(user.password, password):
+#             flash("Wrong Password, Try Again.")
+#             return render_template("/app/templates/accounts/login.html", form=form)
+#         login_user(user)
+#         return redirect(url_for("get_all_posts"))
+#     return render_template("login.html", form=form)
+#
+#
+# @app.route('/logout')
+# @login_required
+# def logout():
+#     logout_user()
+#     return redirect(url_for('get_all_posts'))
 
 
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', port=5000, debug=True)
+    app.run()
